# api/openapi.yaml
openapi: 3.0.3
info:
  title: Ledger API
  version: 0.3.0
  description: |
    API for a lightweight double-entry accounting ledger.

    ## Money
    Every monetary value is an **integer number of cents**. Floating-point
    numbers are never used. Integers are capped at 9007199254740991 (2^53 - 1)
    so they are exact in JavaScript.

    ## Currencies (CAD and USD)
    Every account has one fixed currency, and each line is in its account's
    currency. A journal entry must balance **within each currency**. Amounts
    in different currencies are never added together: reports give totals per
    currency. To move value between currencies, post through an FX clearing
    account in each currency, e.g. buying USD 1,000.00 for CAD 1,370.00:

    | account               | debit    | credit   |
    |-----------------------|----------|----------|
    | 1900 FX clearing CAD  | 1370.00  |          |
    | 1000 Bank CAD         |          | 1370.00  |
    | 1010 Bank USD         | 1000.00  |          |
    | 1910 FX clearing USD  |          | 1000.00  |

    ## Accounts are identified by code
    Accounts are referenced everywhere by their `code` (e.g. `1000`), which
    never changes once created. Internal database ids are not exposed.

    ## Double entry
    A journal entry has at least two lines. Each line is either a debit or a
    credit: exactly one of `debit_cents` / `credit_cents` is greater than zero
    and the other is zero. The database enforces these rules as well.

...

...
servers:
...
paths:
...
  /currencies:
    get:
      tags: [System]
      operationId: listCurrencies
      summary: List supported currencies
      responses:
        '200':
          description: Currencies ordered by code
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CurrencyInfo' }
        default: { $ref: '#/components/responses/Error' }
  /accounts:
    get:
      tags: [Accounts]
      operationId: listAccounts
      summary: List the chart of accounts
      parameters:
        - name: include_inactive
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Accounts ordered by code
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Account' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Accounts]
      operationId: createAccount
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountCreate' }
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /accounts/{accountCode}:
    parameters:
      - $ref: '#/components/parameters/AccountCode'
    get:
      tags: [Accounts]
      operationId: getAccount
      summary: Get one account with its current balance
      responses:
        '200':
          description: The account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        '404': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
    patch:
      tags: [Accounts]
      operationId: updateAccount
      summary: Rename, re-describe or (de)activate an account
      description: |
        The account `code`, `type` and `currency` are not editable. Inactive accounts
        keep their history but cannot receive new postings.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountUpdate' }
      responses:
        '200':
          description: Updated account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /accounts/{accountCode}/ledger:
    parameters:
      - $ref: '#/components/parameters/AccountCode'
    get:
      tags: [Accounts]
      operationId: getAccountLedger
      summary: Posting history of one account with running balance (in the account's currency)
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
      responses:
        '200':
          description: Ledger lines ordered by date, entry id, line number
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountLedger' }
        '404': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /journal-entries:
    get:
      tags: [Journal Entries]
      operationId: listJournalEntries
      summary: List posted journal entries (newest first)
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - name: account_code
          in: query
          description: Only entries with at least one line on this account
          schema: { $ref: '#/components/schemas/AccountCode' }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of journal entries
          content:
            application/json:
              schema:
                type: object
                required: [data, limit, offset]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/JournalEntry' }
                  limit: { type: integer }
                  offset: { type: integer }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Journal Entries]
      operationId: createJournalEntry
      summary: Post a balanced journal entry
      description: |
        The header and all lines are written in a single database transaction.
        If any line is invalid, or the entry does not balance in every
        currency, nothing is saved.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JournalEntryCreate' }
            example:
              entry_date: '2026-10-01'
              description: Office supplies paid in cash
              reference: RCPT-0042
              lines:
                - { account_code: '6100', debit_cents: 4599, credit_cents: 0, memo: Printer paper }
                - { account_code: '1000', debit_cents: 0, credit_cents: 4599 }
      responses:
        '201':
          description: Entry posted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JournalEntry' }
        '422': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /journal-entries/{entryId}:
    parameters:
      - $ref: '#/components/parameters/EntryId'
    get:
      tags: [Journal Entries]
      operationId: getJournalEntry
      summary: Get one journal entry with its lines
      responses:
        '200':
          description: The journal entry
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JournalEntry' }
        '404': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /journal-entries/{entryId}/reverse:
    parameters:
      - $ref: '#/components/parameters/EntryId'
    post:
      tags: [Journal Entries]
      operationId: reverseJournalEntry
      summary: Post a reversing entry (swaps every debit and credit)
      description: An entry can be reversed only once, and a reversal cannot itself be reversed.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReverseRequest' }
      responses:
        '201':
          description: Reversing entry posted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JournalEntry' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /reports/trial-balance:
    get:
      tags: [Reports]
      operationId: getTrialBalance
      summary: Debit/credit balance of every account as of a date, totalled per currency
      parameters:
        - name: as_of
          in: query
          description: Include entries dated on or before this day (default today)
          schema: { type: string, format: date }
        - name: currency
          in: query
          description: Only accounts in this currency
          schema: { $ref: '#/components/schemas/Currency' }
      responses:
        '200':
          description: Trial balance
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrialBalance' }
        default: { $ref: '#/components/responses/Error' }
  /exports/general-ledger:
    post:
      tags: [Files]
      operationId: exportGeneralLedger
      summary: Write the General Ledger to a CSV file in the CSV directory
      description: |
        Columns: `account_code,account_name,account_type,currency,entry_date,
        journal_entry_id,line_id,line_number,reference,description,memo,debit,
        credit,running_balance`. Rows are grouped by account; the running
        balance is per account (debit-positive) in that account's currency.
        When `from` is given, each account starts with an "Opening balance" row.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExportRequest' }
      responses:
        '201':
          description: File written
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CsvFile' }
        '422': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /files:
    get:
      tags: [Files]
      operationId: listCsvFiles
      summary: List CSV files in the CSV directory
      responses:
        '200':
          description: Files, newest first
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CsvFile' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Files]
      operationId: uploadCsvFile
      summary: Upload a CSV file into the CSV directory
      description: |
        The file is stored under its own name, which must match the
        `CsvFilename` pattern. Existing files are never overwritten (409).
        Max 10 MiB; content must be UTF-8 text (a leading BOM is allowed).
        Uploading does not import anything: call `POST /imports/journal-entries`
        afterwards.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: File stored
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CsvFile' }
        '409': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /files/{filename}:
    parameters:
      - $ref: '#/components/parameters/Filename'
    get:
      tags: [Files]
      operationId: downloadCsvFile
      summary: Download a CSV file
      responses:
        '200':
          description: File contents
          content:
            text/csv:
              schema: { type: string, format: binary }
        '404': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /imports/journal-entries:
    post:
      tags: [Files]
      operationId: importJournalEntries
      summary: Import journal entries from a CSV file in the CSV directory
      description: |
        The whole file is imported in **one database transaction**: if any
        entry is invalid, nothing is imported and every problem is reported
        with its row number. A file whose exact content (SHA-256) was imported
        before is refused with 409, so the same file can't be posted twice.

        Format: UTF-8 (a leading BOM is tolerated), comma separated, header
        row required, column names case-insensitive and in any order, at most
        10 MiB and 50,000 rows.

        | column         | required | notes |
        |----------------|----------|-------|
        | `entry_key`    | yes | Rows with the same key form one journal entry; they must be contiguous |
        | `entry_date`   | yes | `YYYY-MM-DD`; may be blank on later rows of an entry |
        | `description`  | yes | May be blank on later rows of an entry |
        | `account_code` | yes | Existing, active account |
        | `debit`        | yes | Decimal, max two decimals (`1234.56`), no signs/separators; blank = 0 |
        | `credit`       | yes | Same as debit; exactly one of debit/credit is non-zero |
        | `reference`    | no  | May be blank on later rows of an entry |
        | `memo`         | no  | Per line |
        | `currency`     | no  | `CAD`/`USD`; if given, must match the account's currency |

        Example:
        ```
        entry_key,entry_date,description,account_code,debit,credit
        S1,2026-10-01,Cash sale,1000,250.00,
        S1,,,4000,,250.00
        ```
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImportRequest' }
      responses:
        '201':
          description: All entries imported
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportResult' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
components:
  parameters:
    AccountCode:
      name: accountCode
      in: path
      required: true
      schema: { $ref: '#/components/schemas/AccountCode' }
    EntryId:
      name: entryId
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
    Filename:
      name: filename
      in: path
      required: true
      schema: { $ref: '#/components/schemas/CsvFilename' }
    From:
      name: from
      in: query
      description: Inclusive start date
      schema: { type: string, format: date }
    To:
      name: to
      in: query
      description: Inclusive end date
      schema: { type: string, format: date }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    AccountType:
      type: string
      enum: [asset, liability, equity, revenue, expense]
    AccountCode:
      type: string
      pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{0,31}$'
      example: '1000'
    Currency:
      type: string
      enum: [CAD, USD]
    CurrencyInfo:
      type: object
      required: [code, name, minor_units]
      properties:
        code: { $ref: '#/components/schemas/Currency' }
        name: { type: string, example: Canadian dollar }
        minor_units: { type: integer, enum: [2] }
...
    Account:
      type: object
      required: [code, name, type, currency, description, is_active, balance_cents, created_at, updated_at]
      properties:
        code: { $ref: '#/components/schemas/AccountCode' }
        name: { type: string }
        type: { $ref: '#/components/schemas/AccountType' }
        currency: { $ref: '#/components/schemas/Currency' }
        description: { type: string }
        is_active: { type: boolean }
        balance_cents:
          allOf: [{ $ref: '#/components/schemas/SignedCents' }]
          description: Current balance in the account's currency (debit-positive)
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AccountCreate:
      type: object
      required: [code, name, type, currency]
      additionalProperties: false
      properties:
        code: { $ref: '#/components/schemas/AccountCode' }
        name: { type: string, minLength: 1, maxLength: 200, example: Bank CAD }
        type: { $ref: '#/components/schemas/AccountType' }
        currency: { $ref: '#/components/schemas/Currency' }
        description: { type: string, maxLength: 1000, default: '' }
    AccountUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 1000 }
        is_active: { type: boolean }
...
    AccountLedger:
      type: object
      required: [account, opening_balance_cents, closing_balance_cents, lines]
      properties:
        account: { $ref: '#/components/schemas/Account' }
        opening_balance_cents: { $ref: '#/components/schemas/SignedCents' }
        closing_balance_cents: { $ref: '#/components/schemas/SignedCents' }
        lines:
          type: array
          items:
            type: object
            required: [journal_entry_id, line_id, entry_date, description, memo, debit_cents, credit_cents, running_balance_cents]
            properties:
              journal_entry_id: { type: integer, format: int64 }
              line_id: { type: integer, format: int64 }
              entry_date: { type: string, format: date }
              description: { type: string }
              reference: { type: string }
              memo: { type: string }
              debit_cents: { $ref: '#/components/schemas/Cents' }
              credit_cents: { $ref: '#/components/schemas/Cents' }
              running_balance_cents: { $ref: '#/components/schemas/SignedCents' }
...